extracting class TarFile from TTC
